Add UserProfilePage with functionality
- Add translations
- Add Route
- Add NavBar link
- Add UserProfile
- Add UserProfileForm
- Add CASCADE to models
- Add editUser functions
- Add deleteUser functions

+++ server/src/interface-adapters/repositories/userRepository.js
@@ -44,9 +44,7 @@ exports.UserRepository = class UserRepository {
   }
 
   async deleteUser(userId) {
-    const result = await User.destroy({
-      where: { userId },
-    });
+    const result = await User.destroy({ where: { userId } });
     return result > 0;
   }
 };
